+++ braintree/customer.py
@@ -159,7 +159,7 @@ class Customer(Resource):
     def create_signature():
         return [
             "company", "email", "fax", "first_name", "id", "last_name", "phone", "website", "device_data", "device_session_id", "fraud_merchant_id", "payment_method_nonce",
-            {"risk_data": ["customer_browser", "customer_ip"]},
+            {"risk_data": ["customer_browser", "customer_device_id", "customer_ip", "customer_location_zip", "customer_tenure"]},
             {"credit_card": CreditCard.create_signature()},
             {"custom_fields": ["__any_key__"]},
             {"three_d_secure_pass_thru": [