+++ tests/unit/test_payment_method_gateway.py
@@ -56,9 +56,10 @@ class TestPaymentMethodGateway(unittest.TestCase):
                                     "first_name",
                                     "last_name",
                                     "locality",
+                                    "phone_number",
                                     "postal_code",
                                     "region",
-                                    "street_address"
+                                    "street_address",
                                 ]
                             },
                         ]